Merge the rd-lookup-base and lookup-type variables in the URI template
The discovery of the individual lookup type implementations is not
hierarchical any more since -10; the separation of the rd-lookup-base
and lookup-type URI variables is a remnant of the old discovery
mechanism.

The distinction between "ep/res mandatory, d/gp optional" was dropped in
this place: The "mandatory"/"optional" discriminator in the variables
lists denote whether the path components are mandatory to be *given*
(which lookup-type always was), while what is mandatory to *implement*
is stated in the {{lookup-types}} table above.

The new unified template variable is type-lookup-location and is defined
in terms of the URI discovery mechanism.