Drop "Base URI" terminology in favor of "URI [path]"
None of the discovered URIs are actually used as "base URIs" (which
would imply constructing URIs relative to it).

Occurrences were replaced with "URI" or "URI paths" as required by the
text flow. Occurrences referring to the LWM2M base URIs were left in
place, as those are actual base URIs.