Drop slashes before URI templates
Both the discovery process and the Location header of an entry or group
creation result in full URIs, even though they are typically expressed
as relative links in .well-known/core, and the Location header is not
even expressive enough to indicate a URI with a different netloc.

The URI templates are changed from `/{+rd}{?...}` to `{+rd}{?...}`, and
the parameter descriptions are updated to not call them "paths" any more
(instead, "locations" where necessary).