Make the recommended </rd> etc. names example values
As suggested in [1] and affirmed in the subsequent group call, the
previously recommended names are now only defaults.

The first lookup example is extended to define all URIs later used, and
gets a label to be referenced.

The "similar to" wording occasionally introduced stems from the
circumstance that the example discovery is run against a multicast
address, but other examples work against a named server.

Link target descriptions of the domain and group lookup interfaces were
left as they are because the example path components there
("/rd/look-up") are a different issue to be resolved.

[1] https://github.com/core-wg/resource-directory/pull/12#pullrequestreview-33531177