Remove more "base" terminology
At some places the term "RD path" was ambiguous (even before the "base"
term was removed, but more visibly so now); after this change, it is
always talked of "RD registration", "RD group" and the four "RD lookup"
URIs or resources.

This is an extension to the 'Drop "Base URI" terminology in favor of
"URI [path]"' commit, but is kept separate for easier review (and should
be smushed before the pull request is merged).