Example addresses: use dedicated subnets
The example addresses previously used a 2001:db8::n:m pattern, where n
indicated the group of examples and m indicated the device, in keeping
with older examples. With that pattern, the distinction between examples
happened in the opaque interface identifier part.

The addresses were changed to a 2001:db8:n::m pattern, where the
examples now form something like independent subnets, so it is clearer
from the addresses that the example groups are independent.

This necessitated the change of the illumination example's RD address
away from :0 (to :ff), as otherwise the RD's address would have collided
with the all-zeros reserved subnet-router anycast address.