Address registered reources by their registration names
This is the most far-reaching of the suggested URI changes.

The endpoint lookup now carries all data that would be needed from a
sketched registration lookup. Moreover, the "contains" relationships
shown in the ER now directly correspond to links, and the properties
(consistently con= as with the registration) denotes attributes of the
target in an extensible way (for n>1).

There is still an open point about how this changes the group
*registration* interface.

In the course of this, the domain lookup interface is dropped (as with
the other commits, this changes the examples first, and if we agree on
how things should look like, we can then describe that).

Closes: #21