5988-vs-6690-anchors: Initial text for siding with 5988
If the WG decides to use the clarified 5988 semantics for href
resolution[1], then this is a starting text.

No attempts have been made here to clean up the terminology, this is
only about the statements about what affects what, and is an update to
the examples.

The last record in the example has been changed to be illustrative with
the new behavior.

[1]: https://mailarchive.ietf.org/arch/msg/core/12u9QBS6q5v5IYknKl_bJYhJKrI