rt is not an attribute of a registration or a group
It's not in the model, is not semantically applicable to the entries
served there (the resource type of a registration entry is something
like "rd-registration", if stated at all).

Curiously, this was explicitly mentioned in the lookup interface
description (without describing its lookup semantics); removing it as
nobody missed it in the model. For resource lookups, this is now
subsumed under resource-param.