lookup: Keep resource-param more generic
The old wording limited lookup to target (or link) attributes -- those
are rt, if, ct etc, but not href, anchor and rel (which are link
parameters but not really target or link attributes).

We've used lookups like href= in examples, so they should also be legal
by the words of the interface.