Make et an independent endpoint parameter
This still defines et, but not in the main body of the specification but
rather like a first example of how additional endpoint parameters should
be done.

It also introduces a registry like Resource Type -- if we don't want
that, we can just drop the registry setup part and move the two
remaining paragraphs a bit up.