Group creation: use the registration resources as in the examples
The examples took up using the registration resources as identifiers in
2155506ef providing a solution to [48], but still passed the endpoint
names too, creating ambiguity in what the server should process; siding
with the target because the endpoint can be dropped then. (And should,
because it serves no purpose; leaving it as SHOULD because no harm comes
from it, and extensions might use it).

The paragraph about the endpoint having to be registered becomes moot
becase when there's no registration, there's no registration resource
either.

Closes: [48]

[48]: https://github.com/core-wg/resource-directory/issues/48