Modify/add JSON and CBOR examples
One examples that was using Link Format on HTTP was changed to use JSON
link format instead (which is expected to be used more in the HTTP
world) where the body was already identical to the previous example.

The simplest lookup example was duplicated and the second version
changed manualy to CBOR Link Format.