Make the registration idempotent at request level
The description of the Registration interface describes it as
idempotent, I think it should be idempotent on CoAP and not only on
application level -- the same request results in the same response with
no further state change.

As to what happens when the re-registration arrives with a *different*
set of links, there we have options; I went for "replaces the
registration" here, but a 4.09 *in that case* would be just as valid an
option.