Forbid Location-Query in registration resources, clean up Location vs.  Location-Path
Location-Query has not been allowed so far but not forbidden either.
This makes a decision against allowing a query component on grounds of
potential conflict with query arguments added later at the
Registration Update operation.

The issue of URI templates that would need modification to do the right
thing (become {&lt,con,...} for registration update rather than
{?lt,con,...} if query is present) additionally weighs in favor of
disallowing them.

Various occurrences of "Location", "Location-Path" and "location"
throughtout the document have been updated to use the correct term for
the HTTP header, the CoAP option and the resulting URI, respectively.

Closes: https://github.com/core-wg/resource-directory/issues/122