Fix misrepresentation of RFC6690's rules
This responds to a comment from Jim on the use of the term "Origin" in
RFC6690 which I had previously missed.

Given that the old 6690-style documents are a lot less powerful, the
text goes from "Submitted documents should be the same in both
interpretations, and the RD can use either" to "Submitted documents
should be the same, and the RD interprets them 8288 style".

A paragraph was added that defines precisely what it means to interpret
it 8288 style.

Several outdated references to where resolution happens have been
removed.