Simplify luminary example
The luminary of the example previously discovered its group addresses by
looking up its own registration resource (which it did not know because
it was entered by a CT), and via that its groups.

As long as the device does not need to know its own registration
resource, the same can be achieved in a single step by looking up groups
that contain it directly, based on the address it would originally have
compared the endpoint results against.

Closes: https://github.com/core-wg/resource-directory/issues/143