"Only information from .well-known/core": insist on should
A "MUST" entered that section in 7ac551239, but I disagree: That's the
conceptional idea, but were it that strict, we'd only have simple
registration. In particular, there can be cases of

* different security concerns applying (eg. "Devcie will never answer to
  .well-known/core because policy says so"), or
* gradual reveal (cf. core-interfaces) happening, where .well-known/core
  lists only a few entry resources, and to get all the information put
  into the RD, a client browsing .well-known/core would need to follow
  links there.

Furthermore, we *do* have some data in the RD that doesn't stem from
.wk/c (and need it for administrative purposes), but we'd like to keep
it to a minimum. A "MUST" here would mean that RD conflicts with its own
design goals.