Remove a reference to the "hosts" relation
This reverts a particular edit from 7ac551239.

The construction of the default base address follows exclusively from
the network address (IP and port) and used protocol. The "hosts" link
relation may, subject to the interpretaton of its very wobbly
definition, also refer to a similar concept, but is not a defining part
of how the base address works.