Explain how the Base URI is constructed
The reference to RFC6690 does not contain text to help in the
construction. Unfortunately, while CoAP has always allowed role reversal
between client and server, TTBOMK there is no text (and not even a term)
to describe "the address that describes the taret of just talking back".

What I've added is the most concise description I could come up with --
I'd prefer to refer to any existing text here.