Demote links-json from normative reference
As links-json went back from the IESG to the WG, chances are it won't be
available as a normative reference, especially given that there's a
tendency to do it right together with CoRAL rather than stay 6690
roundtrippable.

The content formats were removed from the interaction templates (and
replaced with more general phrasing on "web link" media types). The
request to RFC editor to replace TBD64 etc ct= values was turned into a
note to the reader. Examples stayed in, but received notes explicit
references to the draft (to indicate that this example is using a WIP
format).