media types: Losen the limit on link-format equivalence
The previous version limited formats to only supporting a *subset* of
link-format, which was probably not intended.

The new version explicitly allows implementing supersets, though using
their capabilities needlessly is discouraged.

New media formats that do not support particular link-format features
are kept allowed for the simple reason that not even link-format
actually provides all the features it advertises (eg. language-tagged
strings), so checking for whether a language is a superset of
link-format in a prescriptive way would require a better description of
link-format which we don't have.