Move endpoint lookup back to lookup
Endpoint lookup was the only remaining item in the management
interfaces, and those five paragraphs barely warrant an own appendix.
(Plus the introduction would need to be overhauled anyway given
031e81c3.

This effectively reverts what was the registration management
("administrative interface") change of -14 -- but there's much less
that'd go in there now than was back then; see PR discussion for
details.

Closes: https://github.com/core-wg/resource-directory/issues/164
Closes: https://github.com/core-wg/resource-directory/issues/163