Phrase "must give base= when behind NAT" more generally
The requirement was neither precise with current technology (requiring
base= even for NAT with endpoint-independent filtering) nor particularly
future-proof, and is replaced with a more general statement.

While the previous paragraph (about ephemeral endpoints) could have been
pulled into the new one, I decided against it as it'd be deeply nested
("(e.g. because of using ephemeral ports (which happens in TCP))") and
would have borne the risk of implementors overlooking it because of
that.

Closes: https://github.com/core-wg/resource-directory/issues/192