State explicitly how to find an RD using DNS-SD
This pulls a very special application from RD-DNS-SD back into this
document (alternatives were of blocking on or deferring to RD-DNS-SD, or
not having DNS-SD discoverability in RD at all) -- as per the discussion
during IETF106[1]

Having a path= is kept optional (with the consequence of not
guaranteeing URI discovery but falling back to URI discovery if no path
is given) as per RFC6763 Section 6.3.

The value of path is given rather informally in line with other
path=-defining entries in the same registry; "indicate the UTF-8
serialization of a path-abempty shaped reference to the resource on the
server" would be a more formal alternative.

[1]: https://mailarchive.ietf.org/arch/msg/core/78LHFFyq9c1_t0-kAmuDKcTzc3c